Split benchmark Malefile target in two

+++ Makefile
@@ -3,7 +3,9 @@ CFLAGS = -Wall
 
 .PHONY: clean benchmarks
 
-benchmarks: bob-linux bob_qemu-x86_64 run.sh
+benchmarks: benchmarks-linux benchmarks-uk
+
+benchmarks-linux: bob-linux run.sh
 	# read one file of 1Go with 1000Ko buffer
 	./run.sh read linux 1000 1000
 	# read one file of 5Go with 1000Ko buffer
@@ -16,6 +18,8 @@ benchmarks: bob-linux bob_qemu-x86_64 run.sh
 	./run.sh multiple-read linux 1000 1 1000
 	# write 1000 files of 1Mo with 1000Ko buffer
 	./run.sh multiple-write linux 1000 1 1000
+
+benchmarks-uk: bob_qemu-x86_64 run.sh
 	# Unikraft + 9pfs
 	# read one file of 1Go with 1000Ko buffer
 	./run.sh read uk 1000 1000
@@ -38,7 +42,7 @@ bob_qemu-x86_64: main.c Kraftfile
 	cp out/bob_qemu-x86_64 bob_qemu-x86_64
 
 clean:
-	rm bob-linux
-	rm bob_qemu-x86_64
+	[ ! -f 'bob-linux' ] || rm bob-linux
+	[ ! -f 'bob_qemu-x86_64' ] || rm bob_qemu-x86_64
 	kraft clean
